Some changes required for win builds and some changes to ITN model by request of OB.
git-svn-id: https://openmalaria.googlecode.com/svn/trunk@1842 aa1ca794-fd11-11dd-ac65-f74f249cfaac

+++ model/WithinHost/CommonWithinHost.cpp
@@ -133,7 +133,7 @@ int CommonWithinHost::countInfections (int& patentInfections) {
 void CommonWithinHost::checkpoint (istream& stream) {
   WithinHostModel::checkpoint (stream);
   (*pkpdModel) & stream;
-  for (size_t i = 0; i < numInfs; ++i) {
+  for (int i = 0; i < numInfs; ++i) {
     infections.push_back (checkpointedInfection (stream));
   }
   assert( numInfs == infections.size() );
